Move config file, fix window undefined due to react-leaftlet and npm run build successfully
