+++ umap-web/src/components/Map/map.tsx
@@ -1,8 +1,7 @@
 "use client";
-import React ,{ useState} from "react";
-import L, { popup } from "leaflet";
-import { MapContainer, Marker, TileLayer , ZoomControl, useMapEvents,Popup,useMap} from "react-leaflet";
-import { WMSTileLayer, LayersControl} from 'react-leaflet';
+import React, { useState } from "react";
+import { MapContainer, Marker, TileLayer, ZoomControl, useMapEvents, Popup, useMap } from "react-leaflet";
+import { WMSTileLayer, LayersControl } from 'react-leaflet';
 const { BaseLayer, Overlay } = LayersControl;
 import './map.css';
 
@@ -11,11 +10,13 @@ function LocationMarkers() {
   const [markers, setMarkers] = useState([]);
   const map = useMapEvents({
     click(e) {
+      // @ts-ignore
       setMarkers(markers => markers.concat([e.latlng]));
       console.log(markers);
     }
   });
 
+  // @ts-ignore
   const removeMarker = (index) => {
     const newMarkers = markers.filter((_, i) => i !== index);
     setMarkers(newMarkers);
@@ -24,48 +25,48 @@ function LocationMarkers() {
 
   return (
     <>
-      {markers.map((marker,idx) => 
-      <Marker 
-        key = {idx} 
-        position={marker} 
-        eventHandlers={
-          {
-            click(){
-              removeMarker(idx);
+      {markers.map((marker, idx) =>
+        <Marker
+          key={idx}
+          position={marker}
+          eventHandlers={
+            {
+              click() {
+                removeMarker(idx);
+              }
             }
           }
-        }
-      />)}
+        />)}
     </>
   );
 }
 
-export default function MapView(){
-    const [center, setCenter] = useState({lat:10.879961,lng:106.810877});
-    const ZOOM_LEVEL = 12;
-    return (
-        <>
-        <MapContainer
-          center ={center}
-          zoom={ZOOM_LEVEL}
-          scrollWheelZoom={true}
-          zoomControl = {false} 
-          style={{height:"100vh",width:"100vw"}}
-        >
-      
+export default function MapView() {
+  const [center, setCenter] = useState({ lat: 10.879961, lng: 106.810877 });
+  const ZOOM_LEVEL = 12;
+  return (
+    <>
+      <MapContainer
+        center={center}
+        zoom={ZOOM_LEVEL}
+        scrollWheelZoom={true}
+        zoomControl={false}
+        style={{ height: "100vh", width: "100vw" }}
+      >
 
-      <LayersControl>
-        <BaseLayer checked name="U-MAP">
-          <WMSTileLayer url="https://umap.dientoan.vn/geoserver/ows?" layers='TVtesting:planet_osm_line'/>
-        </BaseLayer>
-        <BaseLayer checked name="OSM">
-          <WMSTileLayer url="https://{s}.tile.openstreetmap.fr/hot/{z}/{x}/{y}.png"/>
-        </BaseLayer>
-      </LayersControl>
+
+        <LayersControl>
+          <BaseLayer checked name="U-MAP">
+            <WMSTileLayer url="https://umap.dientoan.vn/geoserver/ows?" layers='TVtesting:planet_osm_line' />
+          </BaseLayer>
+          <BaseLayer checked name="OSM">
+            <WMSTileLayer url="https://{s}.tile.openstreetmap.fr/hot/{z}/{x}/{y}.png" />
+          </BaseLayer>
+        </LayersControl>
 
         <ZoomControl position="topright" />
-        <LocationMarkers/>
-        </MapContainer>
-        </>
-      );
+        <LocationMarkers />
+      </MapContainer>
+    </>
+  );
 }
